wow_server/pyproject.toml
[tool.poetry]
name = "wow-scraping"
version = "0.1.0"
description = ""
authors = ["Christopher Mckerracher <chrislmckerracher@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
typing-extensions = "^4.11.0"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
requests = "^2.31.0"
sqlalchemy = "^2.0.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

